Started migrating to new OOP style.
